requirements.txt
torch
pytorch-lightning
hydra-core
omegaconf
huggingface_hub
pandas
numpy
lmdb
tqdm
wandb
einops
kagglehub
matplotlib
seaborn
scikit-learn
